BE/requirements.txt
# AI Agent - Code Generation & Review API
# Requirements file for Python dependencies

# Web Framework - FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Google Gemini AI
google-generativeai==0.3.2

# Environment Variables
python-dotenv==1.0.0

# Database
pymongo==4.6.0

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2

# Code Quality
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# HTTP Client for testing
requests==2.31.0
